Pick the cleaner version with slurm API.
This is the simplest and more functional version that uses the slurm api
to get the hostname list.
The manual parser was deleted and some extra defensive code checks that
the list is correct for our purposes.

TODO: The extra check needs to be done also then starting the app with
an initial size bigger than one and broadcast the information about the
hosts in use. Such information is required when sing more than one
process/node.