Simplify polling services race condition issue.
The pause function needs to receive all the pendings BUT not send
anything until the new communicator is set in order to avoid sending
more messages in the old communicator.

This commit also simplifies some of the loops in the polling services to
simplify conditions and detect errors.
* src/cluster/polling-services/ClusterWorker.hpp (executeService) : Use
while instead of do-while
* src/cluster/polling-services/MessageDelivery.hpp (waitUntilFinished) :
Restore the execute branch when not done.
* src/cluster/polling-services/MessageHandler.hpp (executeService) : Use
while instead of do-while and simplify condition.
*
src/cluster/polling-services/ClusterServicesPolling.hpp (setPauseStatus)
: waitUntilFinished only for message handler to receive all.
(waitUntilFinished) : Call the new function in MessageHandler.