Fix shrinking and avoid using MPI_Comm_split.
This is the previous step before the generalized version for shrinking.
MPI Messenger now stores the inter and intra communicators as result of
a Spawn in a vector.  When shrinking we can Release the INTRA
communicator and reassign it to the one in the vector back().  The new
code also partially generalizes the case when removing multiple nodes
result of different spawns.
Spawned processes introduce a copy of the MPI_COMM_WORLD and the PARENT
in the first field of the vector so the disconnection logic will be the
always the same without a labyrinth of if-else conditions.