Generalize the shrinking for when removing incomplete groups.
We need to use MPI_Comm_Split when removing processes that where added
as part of a bigger group. This is needed due to a consistency
requirements in the INTRA_COMM.

The current code first iterates over all the communicators that were
spawned after the current one to free the INTRA and disconnect the
inter. When the group is incomplete then the last INTRA_COMM needs to be
split, the killed processes will be in a different INTRA_COMM with the
same parent, so the same code for complete groups can be executed once
again to free INTRA and disconnect INTER (in this case the parent
inter).

The limitation with this approach is that the processes removed won't be
really released until the other processes in the group finalize. But we
can go around this latter doing the spawning by steps/physical nodes.