Add code to share dmalloc info.
After many tests this seems to be the most effective method without too
much complications. The main issue to be aware on this is:

1) The iterative expansion: when spawning on multiple hosts the spawn
can't be made in a single step, but the runtime contains a loop in
nanos6Spawn.

2) There is a synchronization at the end of the dmalloc handling that
needs to be ignored during spawn.

The actual implementation adds a different message type DMALLOC_INFO
which shares all the code with DMALLOC.

* src/cluster/messages/MessageType.hpp : Add new message DMALLOC_INFO.
* src/cluster/polling-services/MessageHandler.hpp : Add new entries for
new message.
* src/cluster/messages/MessageDmalloc.[hc]pp (MessageDmalloc): Receives a
const argument
* src/cluster/ClusterMemoryManagement.cpp : Condition call of
ClusterManager::synchronizeAll based on getType == DMALLOC.
* src/cluster/ClusterManager.[hc]pp (nanos6Resize) : Shares a new
message msgDmallocInfo with new processes.