Add malleability code to be handled by namespace task
The new category of tasks is intended to be handled by the namespace as
"Action Tasks" Which may include taskfinish, and resize (spawn and
shrink) messages.

This is required to solve many nasty issues with malleability and
polling services. So we can completely stop the polling services now;
This was not possible before because the handle of resize message was
executed from the message handler itself which introduced extra
complexity as it was not possible to fully stop them. This change
simplifies and effectively solves all those issues, cleans all the
spagetti code around and also solves many other small issues.

* Makefile.am :
* src/cluster/messages/MessageResize.cpp ->
src/cluster/messages/MessageResizeImplementation.hpp : To implement the
handler functions for resize messages without repeating myself.
* src/cluster/ClusterManager.cpp (nanos6Spawn) :
(nanos6Shrink) : Replaced with handleResizeMessage. Also replaces the
calls to pause the polling servces with really stops and register again.
* src/cluster/ClusterManager.hpp : CHange some functions names and
remove some includes to reduce compilation and cyclic dependencies.