Upgrade the Slurm API to handle hostname order and count.
Add also modifications to get the hostname on every node and share wuth
master when malleability is enabled. This was in the TODOlist, but a
recent upgrade in slurm urged this change because the previous simpler
code was not working anymore.

This change also includes the change to not reset MessageId,
WriteIDManager and OffloadedTaskIdManager after resize.  This produced
issues with the WriteID, so it is better to allocate the maximum on the
beginning and then make it simpler. Remove the resize functions.

* src/cluster/ClusterManager.cpp (ClusterManager) : Get, store and share
the hostname.
(ClusterManager) : Use std::numeric_limits<int>::max() for the init
message tag.
* src/cluster/SlurmAPI.hpp (SlurmHostInfo::changeProcesses) :
(operator<<) :
(getHostinfoByHostname) :
(deltaProcessToHostname) :New functions
(initialize) : Change prototype and reorder hostInfo vector. Also
perform the initial count of already existing processes in every host.

TODO: Add functions to modify these values when malleability.
TODO: Affinity info may be probably stored too.