Remove pause from polling services.
As we can handle the resize messages in the namespace now, we don't need
the complicated and dangerous pause functions for the cluster polling
services.

* src/cluster/ClusterManager.cpp (_pausedServices) :
(_pausedServices) : Remove the static pause variables in
the polling.
Condition the services shutdown because on shrinking they may be already
stopped.
* src/cluster/polling-services/ClusterServicesPolling.hpp :
* src/cluster/polling-services/ClusterServicesTask.hpp : Removed all the
references to the pause variables.
(count) : New function.
(setPauseStatus) : Removed.