Fix WriteID issue with malleability.
When a new node is spawned the WriteIDManager starts the counter from
scratch, this means that after a shrink and another spawn the same
sequence of writeIDs can be regenerated, which is a problem because some
of the previous IDs may be still alive in the WiteIDManagers of nodes
that were alive all the time.

We could delete all the cache infos, but that imply that after a resize
a many redundant data transfers will take place increasing the cost of
a resize with no reason.

In this commit we solve this issue by assigning a new WriteID to any
data with a WriteID created in one of the "shrinking" nodes. Then we
share that information with all the nodes before shrink:

1) the "fetching nodes" can then resister all the transferred region
with the new WriteID.

2) the nodes with copies of that data, can register the new IDs in
their WriteID cache without any information lost.

3) A partial garbage collection for dead WriteIDs can be
performed (cleaning all the IDs Generated in the "shrinking" nodes.

This solves the correctness issue in the most efficient way.

* src/cluster/WriteID.hpp (WriteIDEntry::registerWriteIDInEntry) :
(WriteIDEntry::processIntersectionsWithWriteIDInEntry) : New member
functions to run over the entries. They don't take the locks themselves
because a function may call them several times.
(WriteIDManager::checkWriteIDLocal) :
(WriteIDManager::registerWriteIDasLocal) : Now use the new member
function in entries.
(WriteIDManager::updateExistingWriteID) : New function to search for all
the existing fragments with a WriteID and register it with a different
WriteID. Different to registerWriteIDasLocal this only registers
existing intersecting fragments, not the missing ones.
(WriteIDManager::limitWriteIDToMaxNodes) : New function to cleanup all
the WriteIDs created in certain nodes; this will be extended in the next
commit.
(WriteIDManager::getWriteIDNode) : Demangle the WriteID to the node number.