Update code in ClusterManager
* src/cluster/ClusterManager.cpp (ClusterManager): Move hostname code outside.
Move all the slurm code to the initialization after the constructor to
use the public functions. instead of private ones.
The new initialization gathers all the remote hostnames in Master. So,
multiple send-receive are performed now.
TODO: The SLURM API initialization may be performed in case
initialization time represents an issue in the future.
(nanos6Resize) : Move all the chec code before the taskwait, so if there
is a failure; the assertion/failIf will jump sooner. This is to
integrate with the new SLURM api methods comming next.