Refact all the slurm API
Add all the functions we need to spawn ans shrink the process and all
the code to get information frooom the server and do all the needed
checks.
(SlurmAPI) : The constructor now calls the API to get job, partition and
config information and initialize the variables we need latter and we
are sure won't change.
(requestHostsForNRanks) : Function to request more hosts and exit... no waiting
(checkAllocationRequest) : Function to get the results form the request.
(getSpawnHostInfoVector) : Up to this point ClusterManager only uses
this one which needed to be implemented in advance because it determined
the rest of the implementation.
(releaseUnusedHosts) : Independent function to release all the hosts
without any MPI process. Here we use the internal slurm API to generate
the lists because the compact list seems to be recommended over a over a
huge string.
(updateEnvironment) : This is a private function to update the
environment after a job resize. This is the same (undocumented) actions
that scontrol do.
(hostListToHostinfoVector) : Helper function (will disappear after some
simplifications.)
(getHostinfoByHostname) : Change prototype.
(failIf) : Wrapper for FatalErrorHandler::failIf that adds the slurm
error calling slurm_strerror.
(SlurmHostInfo) Add constructor for emplace (will disappear)