Send dmalloc info earlier.
There was a race condition because the dmalloc info is non blocking, so
it could arrive after some other messages (or being processed after. As
we need that all the new processes have all the memory information, then
we just send that info earlier during the expansion one by one... It may
be less efficient for a huge number of nodes, but in our case it is not
a problem for now.