Fix granularity in process spawn and shrinking.
This basically solves the granularity issue when shrinking processes and
now we don't have the size restriction anymore.

This was intended also to solve the other issue associated with srun
being called in an excessive modified session... There is a bug
apparently in the slurm api that after a limit of modifications a
session is unable to do more srun calls. Such issue is exposed as a
hang in the MPI_Comm_spawn that never returns and there is very few we
can do about it.

All tests are passing here independently form the session allocation.