Add extra field to DmallocDataInfo
This is a field to remember the size when the allocation was
created (may be useful for some optimizations and policies related to
redistribution.

* src/cluster/ClusterMemoryManagement.hpp :
(DmallocDataInfo._clusterInitialSize) : New field
(getSize) : Simplified to avoid errors.
(operator<<) : Added to DmallocDataInfo.
*  src/cluster/ClusterMemoryManagement.cpp (redistributeDmallocs) :
Updates the old clustewrSize field.
* src/cluster/messages/MessageDmalloc.cpp (MessageDmalloc) : Use memcopy
simpler and saver in this case.
* src/memory/directory/cluster/DistributionPolicy.cpp : The cluster size
is now taken form the DmallocDataInfo, not make it an special parameter
or mismatches will occur..