Add new transfer policy: getEagerlyTransfers
This policy transfers all the accesses to their home node.
The function is very different to the lazy policy because it considers
to fragment the accesses if necessary so access remove and reinsert is
necesary.

Probably the insertion and deletion may be implemented using some
processfragment function and similes, but I don't have time to look into
that; so the current version just removes the access and then inserts
the new fragments.