Fix eager transfer policy.
The shrinkInfo must include not only the regions/accesses that will be
transferred, but also should include all the regions and is the receiver
side the one which may decide that to do with all the information.

So the current code fixes that issue and the transfer counter only to
take into account the number of effective transfers which was
wrongly counted before.