Fix policies for the local memory issue.
We need to update the writeID in the eager policy when the accesses are
fragmented in order to update location of read only accesses. This makes
the eager policy to be slower, but this is how it needs to be to work as
intended. When an access is generated before a resize (spawn) and the
latter accesses are only read, then the master node thinks it always own
the region, this may be incorrect after some tasks has executed.

Also the writeid issue we detected some time ago (reusing write id after
a shrink and an spawn (are also aplicable to local memory, s the
exclusion conditions before were wrong and we need to check if the
regions have an ID generated in a dying node unconditionally.

ALL the regions need also to be send for consistency and future
optimizations.