Propagate the noflush at the end of the taskwait.
When taskwait finalizes the accesses were de-fragmented, and contiguous
accesses are merged based on the writeid. This implies that read
accesses require redundant transfers after the operation in the first
task after the taskwait, which is undesirable.
In this commit we inhibit the defragmentation for noflush taskwaits, but
a better fix may be needed that cover also the normal taskwait.