DataLinkStep: don't hold lock when releasing successor steps
DataLinkStep::start used to hold the DataLinkStep's lock when releasing the
successor workflow steps. This had two problems:

(1) Releasing the successor steps may do a lot of work, e.g. starting the
execution step and offloading the task, meaning that the lock could be held for
a long time while it was not needed, delaying other threads.

(2) It risks deadlock due to a lock ordering inversion. If the next workflow
step does something that results in taking a lock on the task's data accesses
in the dependency system, then the overall effect is that this thread first
takes the lock on the DataLinkStep and then on the task's data accesses.
Another thread could take a lock on the task's data accesses to perform an
UpdateOperation (e.g. due to a predecessor task completing), which then calls
DataLinkStep::linkRegion, which in turn takes a lock on the task's
DataLinkStep. The two threads take the locks in opposite orders, giving a
deadlock.