Remove redundant workflow ordering when no copies
When there are no data copies, a redundant ordering was added between the
execution and notification steps. This was probably needed in the past when
(presumably) there was a separate data release step per access. But now that
there is always a single data release step, covering all accesses, the ordering
from execution to notification step is implied by the ordering: execution =>
data release => notification. The explicit execution => notification ordering
was only needed in the past for tasks without accesses, which would have not
had any data release step, so would not have had any link from execution to
notification. Without the explanation here, the previous code was a bit
confusing.