Some compatibility fixes:
This commit completes the Omar's issue when disabling Cluster messenger
in applications.

* src/cluster/ClusterManager.hpp (isSpawned) : This shouldn't trigger
the assertion on Cluster-Disabled applications when called, just return
false.
* src/cluster/ClusterManager.cpp (ClusterManager::ClusterManager) : Set
default values to dataInit limit numbers in order to keep the default -1
for defensive assertions in Cluster-Disabled applications.
* src/cluster/ClusterManagerResizeFunctions.cpp : Simplify the
conditions to really use the default values and avoid mismatches between
the message and the real behavior.
* src/cluster/ClusterUtil.hpp : C++ does not cast anymore the strings to
char*, so explicitly use c_str.