Cluster node reduction location and satisfiability tracking
Since reductions are combined at the original node, there is no need
to pass the true read and write satisfiability to the remote node. Always
pass read and write satisfiability immediately. The original value is
never sent to the remote node, so give it a location of -42, meaning that
it is uninitialized (from the point of view of the remote node).