Offload task immediately if destination rank is known
The basic idea is to convert all accesses of the local task into weak accesses.
This will allow the local task to be executed immediately, so that it can
create the local workflow and offload the task. Normal tasks have weak accesses
because they don't access the data - only their subtasks do. Similarly, the
local copy of a task that will be offloaded doesn't access the data - only
the task it creates (the copy of the task on the remote rank) does.

The change was surprisingly easy. When the task is submitted, after registering
the task's accesses, we check whether it will be offloaded. The code currently
only works with the "node" clause, but a "pre-allocation" step could inspect
the accesses and decide which rank the task should be executed on. If the task
will be offloaded, all in, inout and out accesses are converted to weak
accesses (but the access keeps a bit to remember that the access was originally
strong).  Note: we need to be careful to do this before the task is released by
checkSubmittedTaskReady (there would be a race condition and potential
use-after-free, noted in the code).