Fix rare hang in manual_empty
During the shutdown process, the cluster manager unregisters the ClusterWorkers
(in ClusterManager::shutdownPhase1). It first checks the number of cluster
workers, using ClusterServicesTask::count(), but an unfortunate race condition
meant that this function used to return zero before the cluster services tasks
had executed for the first time.  If the main task finishes very quickly (e.g.
in manual_empty), then, since it appears that there are no cluster workers,
they aren't unregistered. These unfinished tasks caused Nanos6 to hang in
ThreadManager::shutdownPhase1. The fix is to make ClusterServicesTask::count()
return the correct value, by incrementing _activeClusterTaskServices as soon as
the service is registered rather than when the task is first executed.