Register namespace predecessor of offloaded tasks only once
remoteTaskCreateAndSubmit passes the global task ID of the task preceding the
current access as an argument of
DataAccessRegistration::registerTaskDataAccess. The dependency system uses this
value to determine whether it is valid to link the access with its predecessor
on the bottom map of the namespace task (if the task IDs don't match it is
because of out-of-order task offloading, an intervening write by a task on a
different node or a task offloaded by a different task to the same node - all
of which mean that linking the accesses in the namespace is invalid).  This ID
is taken from the _id field of the SatisfiabilityInfo passed in the TaskNew
message. Later, when propagating the read and write satisfiabilities embedded
in the TaskNew message, remoteTaskCreateAndSubmit unnecessarily passed the
namespace predecessor to propagateSatisfiability, giving it to the dependency
system for a second time.  This was done even when the access of the newly
submitted task was not yet read or write satisfied, so no call to
propagateSatisfiability was really necessary.  The result was some unnecessary
traversals of the task data accesses.