Remove disableReadPropagationToNext from removeBottomMapTaskwaitOrTopLevelSink
There used to be an "optimization" that delayed the propagation of read
satisfiability when the data is being fetched, until the data arrives and the
location is local. This "optimization" was misguided and removed by "Disable
disableReadPropagationToNext". The commit message explains why: essentially it
delayed all the successor tasks that would execute on another node so don't
need to wait for the data to be copied to the original node. The problem was
better solved with a more effective implementation of WriteID. Here, remove the
special handling for disableReadPropagationToNext inside
removeBottomMapTaskwaitOrTopLevelSink.