Data copy register location immediately not requiring a later pass
If a strong access doesn't need a data transfer, then it is only necessary to
update its location (e.g. the access is an "out", the data is as-yet
uninitialized, various other reasons in the calculation of needsTransfer, or it
was found to be locally present by WriteID).  Since the data copy step is
created with the lock on the task accesses and a pointer to the actual access,
we can update the access' location immediately.  This saves a traversal of the
task accesses that would otherwise have been required when doing it later.