Taskwait: no output location for dmalloc'ed or read-only data
Data that was dmalloc'ed by the current task or that is read-only for the
current task never needs to be fetched by a #pragma taskwait. In this case
createTaskwait used to set the output location to the current node, but then
the data copy step discovered that it didn't need a transfer and set the output
location back to nullptr. This was OK, but confusing. It is better to never set
the output location in this case.