Quick path for fragmentAccess in common case
Fragmenting an access used to go through the whole DataAccessStatusEffects
rigamarole. But most of the time there is nothing to do apart from updating the
removalBlockers and liveTaskwaitFragmentCount (where necessary). Moreover the
access is always already registered and usually not yet removable. A quick
check can identify this common case and use special purpose code. The only
downside is some duplication between handleDataAccessStatusChanges and
setUpNewFragment, which is made clear in the comments. Fragmentation can be a
non-avoidable sequential bottleneck, so this duplication is worth tolerating.