Move check that node() is in range into AddTask
An error: "node in node() clause is out of range [...]" is generated if the
argument to the node() clause is an out-of-range number. This used to be
checked by the scheduling policy, but it is better to do it as soon as possible
in AddTask. This allows later parts of the code to assume that any non-negative
node constraint is a valid cluster node ID.