Fix a comment regarding namespace propagation into ins
A comment in applyUpdateOperationOnAccess used to claim that in=>in and
non-in=>in propagation in the namespace passed R and C satisfiability in the
namespace and W satisfiability via the propagateSatisfiability message from the
offloader. In fact, C satisfiability was never passed in this situation. We
update the comment and code to pass C satisfiability with the
propagateSatisfiability message. This is the correct thing to do, because
passing it in the namespace would imply (1) an extra and unnecessary dependency
operation when the previous task finishes, just to pass C satisfiability and
(2) this redundant operation may happen after the successor task has received W
satisfiability from the offloader (as well as R in the namespace), so the task
could have already been deleted (C satisfiability is not part of the
calculation of _isRemovable). Hence the approach outlined in the comment risked
a use-after-free error.