Reduction data copy original value for offloaded task
The first access in a reduction is responsible for copying in the original data
value, if needed. If the access is already read satisfied when the workflow is
created, then the data copy step is created immediately. If it is not yet read
satisfied, then the data copy step will be created later when it becomes read
satisfied (this is done through bytesReductionLaterDataCopies and the
CounterStep). Unfortunately if this first access belongs to an offloaded task,
then the "data copy" step created by the workflow was actually a data link
step. In this case, it is necessary to create another data copy step to
actually copy the original value. This fixes reductions_reductions22.