Add --with-tampi=<prefix> to configure
When using OmpSs-2@Cluster with TAMPI, the runtime will need to call the TAMPI
API (a) to control the initialization process and (b) to turn off TAMPI task
awareness for the few blocking MPI calls issued by the runtime. So it needs to
be built with a new --with-tampi configure option. Note: TAMPI and the runtime
can be linked in either order.