Add variant of ClusterMemoryManagement::lfree without size
Update lmalloc to record the size of the allocated region, by allocating an
extra cache line at the beginning to hold this information. Update the existing
variant of lfree, which has a parameter giving the size of the allocation, to
check that the value from the user is correct. Add a variant of lfree without
the size parameter, which just reads the value recorded by lmalloc.

There are two complexities:

(1) It is valid to do the lfree on a different node than the one that did the
allocation. An example is nasty/test3_hang3.c in the unit tests. In this case,
the node doing the lfree won't have a copy of the allocated size, so it is not
possible to check the value passed by the user. This also means that the
variant without the size parameter cannot be used.

(2) NUMAManager::alloc abuses the interface by freeing only part of the
allocated region. In this case it is also not possible to check the size.