NUMAManager::alloc rounds up to next page
It used to assert that the allocated region is aligned on a page boundary.  But
this only works if all previous lmallocs had size a multiple of a page, so it
was fragile. This fixes it, but unfortunately it does call a memory leak with
total size one page per NUMA allocation. This should be fixed in the future,
but is unlikely to be a major issue as NUMA allocations should not be very
frequent in real programs.