Error if one core and cpumanager.reserve_cpu_leader=true
When cpumanager.reserve_cpu_leader = true and the process has only one core,
there are no cores available to run any tasks (which results in a hang). Raise
an error if this is the case. A change in the Slurm configuration on
MareNostrum means that this happens when using mpirun.