Add variant of ClusterMemoryManagement::dfree without size
The correct value of the 'size' argument can be deduced from the _dmallocs list
that is already held by the ClusterMemoryManagent. This list, of all the
current dmallocs, is present on all nodes. It exists so that all nodes can
update the affinity of dmallocs on a malleability operation. In this context, a
malleability operation is either a change in the number of processes (process
spawn/shrink) or a change in the number of cores of a process (with DLB
integration: different nodes may have an inconsistent view, but this is OK).
Neither of these operations are relevant to a dfree, but the list of dmallocs
can be used to look up the size of a dmalloc given its start address. Update
the normal dfree method, with size argument, to check that the size is correct.
Add an alternative dfree method with the start address but no size.  Since the
list of dmallocs is duplicated on all nodes, both variants can be initiated on
any node, not necessarily the one that did the dmalloc.